+++ tests/common.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <libtcod/console.h>
+
 #include <cstdlib>
 #include <string>
 
@@ -10,4 +12,20 @@ static inline std::string get_file(const std::string& path) {
     data_dir = DEFAULT_DIR;
   }
   return std::string(data_dir) + "/" + path;
+}
+
+/*****************************************************************************
+    @brief Convert a console to a multi-line string, used for tests.
+ */
+static inline std::string to_string(const TCOD_Console& console) {
+  std::string result;
+  for (int y = 0; y < console.h; ++y) {
+    if (y != 0) {
+      result += '\n';
+    }
+    for (int x = 0; x < console.w; ++x) {
+      result += static_cast<char>(console.at(x, y).ch);
+    }
+  }
+  return result;
 }