+++ CHANGELOG.md
@@ -13,6 +13,7 @@ since `1.7.0`.
 - Several C++ drawing and print functions have been made public.
 - Several C++ tileset functions have been made public.
 - C++ context functions are now public.
+- Added vprintf-like console printing functions.
 
 ### Deprecated
 - Deprecate `TCOD_chars_t` enum values because they are non-Unicode.